+++ src/index.ts
@@ -1,4 +1,3 @@
 import DatasetJson from './class/datasetJson';
 export default DatasetJson;
 export * from './interfaces/datasetJson';
-export * from './interfaces/filter';