+++ test/filters.test.ts
@@ -1,18 +1,20 @@
 import DatasetJson from '../src/index';
-import { Filter } from '../src/interfaces/filter';
+import Filter from 'js-array-filter';
 
 test('Get filtered rows of dataset with simple "and" filter', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'gt', value: 80 },
             { variable: 'SEX', operator: 'eq', value: 'M' }
         ],
         connectors: ['and']
-    };
-    const rows = await data.getData({ start: 0, length: 5, filterData: filter, filterColumns: ['USUBJID', 'SEX', 'AGE'] });
+    });
+    const rows = await data.getData({ start: 0, length: 5, filter, filterColumns: ['USUBJID', 'SEX', 'AGE'] });
     expect(rows.length).toBeLessThanOrEqual(5);
     expect(rows).toMatchSnapshot();
 });
@@ -21,14 +23,16 @@ test('Get filtered rows of dataset with simple "or" filter', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'gt', value: 85 },
             { variable: 'DCDECOD', operator: 'eq', value: 'STUDY TERMINATED BY SPONSOR' }
         ],
         connectors: ['or']
-    };
-    const rows = await data.getData({ start: 0, filterData: filter, filterColumns: ['USUBJID', 'DCDECOD', 'AGE'] });
+    });
+    const rows = await data.getData({ start: 0, filter, filterColumns: ['USUBJID', 'DCDECOD', 'AGE'] });
     expect(rows.length).toEqual(24);
 });
 
@@ -36,15 +40,17 @@ test('Get filtered rows of dataset with eq operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'DCDECOD', operator: 'eq', value: 'STUDY TERMINATED BY SPONSOR' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'DCDECOD']
     });
     expect(rows.length).toEqual(7);
@@ -54,15 +60,17 @@ test('Get filtered rows of dataset with contains operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'RACE', operator: 'contains', value: 'WHITE' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'RACE']
     });
     expect(rows.length).toEqual(230);
@@ -72,16 +80,18 @@ test('Get filtered rows of dataset with contains operator and case insensitive o
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'RACE', operator: 'contains', value: 'bLACK' }
         ],
         connectors: [],
         options: { caseInsensitive: true }
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'RACE']
     });
     expect(rows.length).toEqual(23);
@@ -91,15 +101,17 @@ test('Get filtered rows of dataset with notcontains operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'RACE', operator: 'notcontains', value: 'WHITE' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'RACE']
     });
     expect(rows.length).toEqual(24);
@@ -109,15 +121,17 @@ test('Get filtered rows of dataset with starts operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'TRT01P', operator: 'starts', value: 'Xanomeline Low' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'TRT01P']
     });
     expect(rows.length).toEqual(84);
@@ -127,15 +141,17 @@ test('Get filtered rows of dataset with ends operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'TRT01P', operator: 'ends', value: 'ebo' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'TRT01P']
     });
     expect(rows.length).toEqual(86);
@@ -145,15 +161,17 @@ test('Get filtered rows of dataset with regex operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'TRT01P', operator: 'regex', value: '^Xano.*Dose$' }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'TRT01P']
     });
     expect(rows.length).toEqual(168);
@@ -163,16 +181,18 @@ test('Get filtered rows of dataset with regex operator and case insensitive opti
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'TRT01P', operator: 'regex', value: '^pLaCEBO$' }
         ],
         connectors: [],
         options: { caseInsensitive: true }
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'TRT01P']
     });
     expect(rows.length).toEqual(86);
@@ -182,15 +202,17 @@ test('Get filtered rows of dataset with in operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'USUBJID', operator: 'in', value: ['01-701-1015', '01-702-1082'] }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID']
     });
     expect(rows.length).toEqual(2);
@@ -200,15 +222,17 @@ test('Get filtered rows of dataset with notin operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'DCDECOD', operator: 'notin', value: ['ADVERSE EVENT', 'DEATH', 'COMPLETED'] }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'DCDECOD']
     });
     expect(rows.length).toEqual(49);
@@ -218,16 +242,18 @@ test('Get filtered rows of dataset with in operator and case insensitive option'
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'DCDECOD', operator: 'in', value: ['ADVERSE event', 'DeAtH', 'ComplETED'] }
         ],
         connectors: [],
         options: { caseInsensitive: true }
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'DCDECOD']
     });
     expect(rows.length).toEqual(205);
@@ -237,15 +263,17 @@ test('Get filtered rows of dataset with gt operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'gt', value: 80 }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'AGE']
     });
     expect(rows.length).toEqual(77);
@@ -255,15 +283,17 @@ test('Get filtered rows of dataset with lt operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'lt', value: 53 }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'AGE']
     });
     expect(rows.length).toEqual(2);
@@ -273,15 +303,17 @@ test('Get filtered rows of dataset with ge operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'ge', value: 89 }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'AGE']
     });
     expect(rows.length).toEqual(1);
@@ -291,15 +323,17 @@ test('Get filtered rows of dataset with le operator', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'le', value: 51 }
         ],
         connectors: []
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
+        filter,
         filterColumns: ['USUBJID', 'AGE']
     });
     expect(rows.length).toEqual(1);
@@ -309,7 +343,9 @@ test('Get filtered rows of dataset with all types of operators', async () => {
     const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const filter: Filter = {
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'DCDECOD', operator: 'eq', value: 'STUDY TERMINATED BY SPONSOR' },
             { variable: 'RACE', operator: 'contains', value: 'BL' },
@@ -326,12 +362,12 @@ test('Get filtered rows of dataset with all types of operators', async () => {
             { variable: 'AGE', operator: 'le', value: 60 },
         ],
         connectors: ['or', 'or', 'and', 'or', 'and', 'or', 'or', 'and', 'or', 'and', 'or', 'or']
-    };
+    });
     const rows = await data.getData({
         start: 0,
-        filterData: filter,
-        filterColumns: ['USUBJID', 'SEX', 'AGE', 'RACE', 'TRT01P', 'DCDECOD', 'DSDECOD'] }
-    );
+        filter,
+        filterColumns: ['USUBJID', 'SEX', 'AGE', 'RACE', 'TRT01P', 'DCDECOD', 'DSDECOD']
+    });
     expect(rows.length).toEqual(75);
     expect(rows).toMatchSnapshot();
 });